pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sp-pipeline"
version = "0.1.0"
description = "Automated pipeline for querying signal peptide data from curated databases"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "SP-Pipeline Contributors"}
]
keywords = ["signal-peptide", "bioinformatics", "proteomics", "uniprot", "ncbi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "requests>=2.31.0",
    "biopython>=1.83",
    "pandas>=2.0.0",
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "platformdirs>=4.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
]

[project.scripts]
sp-pipeline = "sp_pipeline.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
